chore: renaming noteDetail
